requirements.txt
windows-curses==2.4.1
python-dotenv==1.2.1